# Windows,Bazel client: check embedded tools faster