--- a/src/main/cpp/blaze.cc
+++ b/src/main/cpp/blaze.cc
@@ -1062,7 +1062,7 @@ static void ExtractData(const string &self_path) {
         globals->options->install_base, "_embedded_binaries");
     for (const auto &it : globals->extracted_binaries) {
       string path = blaze_util::JoinPath(real_install_dir, it);
-      if (!mtime.get()->IsValidEmbeddedBinary(path)) {
+      if (!mtime.get()->IsUntampered(path)) {
         BAZEL_DIE(blaze_exit_code::LOCAL_ENVIRONMENTAL_ERROR)
             << "Error: corrupt installation: file '" << path
             << "' modified.  Please remove '" << globals->options->install_base