--- a/src/main/cpp/util/file_platform.h
+++ b/src/main/cpp/util/file_platform.h
@@ -31,11 +31,15 @@ class IFileMtime {
  public:
   virtual ~IFileMtime() {}
 
-  // Checks if `path` is a valid embedded binary, i.e. it is either an
-  // embedded binary that Bazel extracted from a previous run and it hasn't
-  // been tampered with (its mtime is unchanged), or it's a directory.
+  // Checks if `path` is a file/directory in the embedded tools directory that
+  // was not tampered with.
+  // Returns true if `path` is a directory or directory symlink, or if `path` is
+  // a file with an mtime in the distant future.
   // Returns false otherwise, or if querying the information failed.
-  virtual bool IsValidEmbeddedBinary(const std::string &path) = 0;
+  // TODO(laszlocsomor): move this function, and with it the whole IFileMtime
+  // class into blaze_util_<platform>.cc, because it is Bazel-specific logic,
+  // not generic file-handling logic.
+  virtual bool IsUntampered(const std::string &path) = 0;
 
   // Sets the mtime of file under `path` to the current time.
   // Returns true if the mtime was changed successfully.