@cvcal 2018-06-21
Did you also time the difference on linux?