@laszlocsomor 2018-06-21
> Did you also time the difference on linux?

I just did. The PR improves that too. Updated the commit message.