@cvcal 2018-06-21
Assuming this is superseded with https://github.com/bazelbuild/bazel/pull/5445