@laszlocsomor 2018-06-21
> Assuming this is superseded with #5448

No, they are separate PRs and will be imported as separate commits.
However GitHub seems to have no good support for something like a Diffbase -- GitHub shows the entire branch as it diverged from `master`, and I can't tell it to display it based on this PR (#5445).